Finish implementing Freight for FreightProxy
While implementing Freight, it was also found that it's better
to return functions, types, modules and traits, when found by name
in a vector of all of them sharing that name.

The name is no longer a unique identifier for anything. ID for
everything and native_id for types retained their uniqueness
and will remain this way. ID is going to remain unique because
of it's nature -- it was made to be something that uniquely
and optimally identifies an item in the list of items, and
native_id will remain this way, because the type representation
in rust only allows one TypeId for one type, which means that
in order for us to be able to distinguish different types
we need TypeId to be unique.